pyproject.toml
[tool.poetry]
name = "fintrendanalyser"
version = "0.1.0"
description = "FinTrendAnalyser is a comprehensive financial data analysis platform"
authors = ["Kirill Penchukov <kirill.penchukov.993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.9.1"
aiosignal = "1.3.1"
alpha-vantage = "2.3.1"
appdirs = "1.4.4"
attrs = "23.1.0"
beautifulsoup4 = "4.12.2"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
click = "8.1.7"
contourpy = "1.2.0"
cycler = "0.12.1"
decorator = "5.1.1"
fonttools = "4.46.0"
frozendict = "2.3.10"
frozenlist = "1.4.0"
html5lib = "1.1"
idna = "3.6"
joblib = "1.3.2"
kiwisolver = "1.4.5"
lxml = "4.9.3"
matplotlib = "3.8.2"
multidict = "6.0.4"
multitasking = "0.0.11"
nltk = "3.8.1"
numpy = "1.26.2"
oauthlib = "3.2.2"
packaging = "23.2"
pandas = "2.1.3"
peewee = "3.17.0"
pillow = "10.1.0"
psycopg2 = "2.9.9"
pyparsing = "3.1.1"
pyrate-limiter = "3.1.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
pytz = "2023.3.post1"
regex = "2023.12.25"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
sec-edgar-downloader = "5.0.2"
six = "1.16.0"
soupsieve = "2.5"
tabulate = "0.9.0"
textblob = "0.17.1"
tqdm = "4.66.1"
tweepy = "4.14.0"
tzdata = "2023.3"
urllib3 = "2.1.0"
wbdata = "0.3.0"
webencodings = "0.5.1"
yarl = "1.9.4"
yfinance = "0.2.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
